pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smarteval"
version = "0.1.0"
description = "A Python-first framework for evaluating non-deterministic systems."
readme = "DESIGN.md"
requires-python = ">=3.12"
dependencies = [
  "jsonschema>=4.22,<5",
  "openai>=1.109.0",
  "pydantic>=2.7,<3",
  "PyYAML>=6.0",
  "typer>=0.12,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
smarteval = "smarteval.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
